# perf(http) reduce memcpy