Moving Conn around caused a large number of memcpy calls to be done.
Instead use a Box and &mut self in place of moves.

Not done yet, fixing up more. Need to resolve the self.on_remove() move
issue.

This took hello.rs from ~400K req/s to close to ~500K req/s for me