@Keats 2016-06-08
400-500k req/s? How are you testing it?

I get ~105k req/s on a good 2 years laptop that way:
`cargo run --example hello --release`
`wrk --latency -t12 -c100 -d20s http://localhost:3000`