@bfrog 2016-06-08
Testing it on my core i7 desktop

```
wrk --latency -c 500 -d 30s -t 12 http://localhost:3000/                                   ✓  6952  11:14:10 
Running 30s test @ http://localhost:3000/
  12 threads and 500 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency     1.30ms    1.68ms  47.35ms   88.47%
    Req/Sec    41.40k    20.85k  170.33k    69.83%
  Latency Distribution
     50%    0.88ms
     75%    1.40ms
     90%    3.20ms
     99%    6.55ms
  13930273 requests in 29.94s, 1.14GB read
Requests/sec: 465263.03
Transfer/sec:     39.05MB
```