@bluetech 2016-06-10
Just a data point, here's `perf record` of the above `wrk` command before this PR (only >= 2%, rust 1.9.0):

```
    17.14%  hello    libc-2.23.so        [.] __memcpy_avx_unaligned
     9.93%  hello    libc-2.23.so        [.] __memset_avx2
     5.64%  hello    hello               [.] hyper::http::h1::parse::_$LT$impl$u20$http..Http1Message$u20$for$u20$http..ServerMessage$GT$::parse::h95f4390be86b6454
     5.26%  hello    hello               [.] _$LT$hyper..http..conn..State$LT$H$C$$u20$T$GT$$GT$::update::hda2a28a5f3d7f4ac
     4.34%  hello    hello               [.] core::fmt::write::h4eb847ed3e152952
     4.00%  hello    hello               [.] _$LT$std..io..Write..write_fmt..Adaptor$LT$$u27$a$C$$u20$T$GT$$u20$as$u20$std..fmt..Write$GT$::write_str::h428ca1799ea8eabb
     3.95%  hello    hello               [.] core::fmt::Formatter::write_fmt::ha8631b2bdb82a0b7
     3.76%  hello    hello               [.] _$LT$hyper..http..conn..Conn$LT$K$C$$u20$T$C$$u20$H$GT$$GT$::on_writable::h150e5d121479d65f
     2.88%  hello    hello               [.] _$LT$mio..event_loop..EventLoop$LT$H$GT$$GT$::run_once::h42b5013245ea60d3
     2.16%  hello    hello               [.] mallocx
     2.07%  hello    hello               [.] sdallocx
     2.07%  hello    hello               [.] time::display::parse_type::h4267bac127e2a6cf
```

and after:

```
     6.57%  hello    hello               [.] _$LT$hyper..http..conn..State$LT$H$C$$u20$T$GT$$GT$::update::hda2a28a5f3d7f4ac
     6.15%  hello    hello               [.] hyper::http::h1::parse::_$LT$impl$u20$http..Http1Message$u20$for$u20$http..ServerMessage$GT$::parse::h95f4390be86b6454
     6.04%  hello    hello               [.] _$LT$mio..event_loop..EventLoop$LT$H$GT$$GT$::run_once::h42b5013245ea60d3
     5.61%  hello    hello               [.] _$LT$std..io..Write..write_fmt..Adaptor$LT$$u27$a$C$$u20$T$GT$$u20$as$u20$std..fmt..Write$GT$::write_str::h428ca1799ea8eabb
     4.92%  hello    hello               [.] core::fmt::write::h4eb847ed3e152952
     4.88%  hello    libc-2.23.so        [.] __memcpy_avx_unaligned
     4.85%  hello    hello               [.] _$LT$hyper..http..conn..Conn$LT$K$C$$u20$T$C$$u20$H$GT$$GT$::on_writable::h150e5d121479d65f
     3.28%  hello    hello               [.] rotor::handler::replacer::ha0a65d97791110f7
     3.28%  hello    hello               [.] core::fmt::Formatter::write_fmt::ha8631b2bdb82a0b7
     3.14%  hello    hello               [.] mallocx
     2.85%  hello    hello               [.] _$LT$hyper..http..conn..Conn$LT$K$C$$u20$T$C$$u20$H$GT$$GT$::ready::h4913e9752920612a
     2.76%  hello    libc-2.23.so        [.] __memset_avx2
     2.64%  hello    hello               [.] time::display::parse_type::h4267bac127e2a6cf
     2.31%  hello    hello               [.] sdallocx
```

So this PR does do what it says on the tin. I guess when the string formatter shows near the top, the situation is not bad :)