@bfrog 2016-06-10
This should be good now minus I guess styling/code likeability. It does what its meant to though.

Before these changes, for the hello example
``` sh
wrk --latency -c 500 -d 30s -t 12 http://localhost:3000/
Running 30s test @ http://localhost:3000/
  12 threads and 500 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency     1.63ms    1.89ms  46.11ms   88.65%
    Req/Sec    38.26k    25.27k  164.33k    70.47%
  Latency Distribution
     50%  804.00us
     75%    2.93ms
     90%    3.86ms
     99%    6.60ms
  12646358 requests in 29.98s, 1.04GB read
Requests/sec: 421892.96
Transfer/sec:     35.41MB
```

After these changes for the hello example
``` sh
wrk --latency -c 500 -d 30s -t 12 http://localhost:3000/
Running 30s test @ http://localhost:3000/
  12 threads and 500 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency     1.11ms    1.26ms  49.31ms   87.87%
    Req/Sec    44.22k    16.81k  153.22k    67.20%
  Latency Distribution
     50%  700.00us
     75%    1.38ms
     90%    2.69ms
     99%    5.77ms
  14946137 requests in 29.95s, 1.22GB read
Requests/sec: 499083.80
Transfer/sec:     41.88MB
```