@Ryman 2016-06-10
There's a lot of `self.0.foo` action, would it make sense to implement `Deref` / `DerefMut` on Conn?